Add conditional calculator of batch size
Previously the dataGenerator required batch size to be supplied as an argument in the terminal.
This was helpful for testing, as it allowed me to easily experiment with different batch
sizes without needing to repeatedly edit my code. Now that I have a good idea of optimal batch
sizes based on save method & entry qty, it makes more sense to calculate the batch size in the
code, rather than ask the user to supply it.